Move battle back into root onix directory
